Refactor bandwidth parser tests into modules
